Added Users and changed Tasks Controller

+++ controllers/tasks.controller.js
@@ -1,15 +1,29 @@
-import Task from "../models/tasks.models.js"
+import Task from "../models/tasks.models.js";
+import User from "../models/user.models.js";
 
 export const getAllTasks = async (req, res) => {
   try {
-    //fetching all the tasks
-    const tasks = await Task.find();
-    //if there are not tasks in the DB
+
+    const { userId } = req.params;
+
+    const user = await User.findById(userId);
+    if (!user) {
+      return res
+        .status(400)
+        .json({ message: "User not Found", success: false });
+    }
+
+    const tasks = await Task.find({ user: userId });
+    //if there are not tasks in DB
     if (tasks.length === 0) {
-  return res.status(404).json({ message: "No Tasks Found", success: false });
-}
-    
-    return res.status(200).json({ message: "All tasks fetched successfully", tasks });
+      return res
+        .status(404)
+        .json({ message: "No Tasks Found", success: false });
+    }
+
+    return res
+      .status(200)
+      .json({ message: "All tasks fetched successfully", tasks });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -17,23 +31,30 @@ export const getAllTasks = async (req, res) => {
 
 export const newTask = async (req, res) => {
   try {
+    const { userId } = req.params;
     //taking taskname and description as inputs
-    const {taskName, taskDescription} = req.body;
+    const { taskName, taskDescription } = req.body;
 
-    //no Need
-    // if(!taskName || !taskDescription){
-    //   return res.status(400).json({message: "Missing Fields", success: false})
-    // }
+    const user = await User.findById(userId);
+    if (!user) {
+      return res
+        .status(400)
+        .json({ message: "User not Found", success: false });
+    }
+
 
     const newTask = new Task({
       taskName,
-      taskDescription
+      taskDescription,
+      user
     });
 
     //saving in DB
     await newTask.save();
 
-    res.status(201).json({ message: "Task created successfully", success: true, newTask});
+    res
+      .status(201)
+      .json({ message: "Task created successfully", success: true, newTask });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -43,14 +64,32 @@ export const getSingleTask = async (req, res) => {
   try {
     //taking taskId from params
     const { id } = req.params;
+    const { userId } = req.params;
+
+    const user = await User.findById(userId);
+    if (!user) {
+      return res
+        .status(400)
+        .json({ message: "User not Found", success: false });
+    }
 
     const task = await Task.findById(id);
     //checking if task is present in DB or no
-    if(!task){
-      return res.json({message: "Task Not Found", success: false})
+    if (!task) {
+      return res.json({ message: "Task Not Found", success: false });
     }
+
+    if (task.user != userId) {
+      return res
+        .status(401)
+        .json({ message: "Task Not Found", success: false });
+    }
+
     
-    return res.status(200).json({message: "Task Found Successfully", success: true, task})
+
+    return res
+      .status(200)
+      .json({ message: "Task Found Successfully", success: true, task });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -58,25 +97,42 @@ export const getSingleTask = async (req, res) => {
 
 export const updateTask = async (req, res) => {
   try {
-    //taking taskId from params
+    //taking taskId and userId from params
     const { id } = req.params;
+    const { userId } = req.params;
 
+    const user = await User.findById(userId);
+    if (!user) {
+      return res
+        .status(400)
+        .json({ message: "User not Found", success: false });
+    }
+    
     const task = await Task.findByIdAndUpdate(id);
     //checking if task is present in DB or no
-    if(!task){
-      return res.json({message: "Task Not Found", success: false})
+    if (!task) {
+      return res.json({ message: "Task Not Found", success: false });
     }
+
+    if (task.user != userId) {
+      return res
+        .status(401)
+        .json({ message: "Task Not Found", success: false });
+    }
+    
     //condition for updating task completed
-    if(task.completed==false){
-      task.completed=true;
+    if (task.completed == false) {
+      task.completed = true;
+    } else {
+      task.completed = false;
     }
-    else{
-      task.completed=false;
-    }
+    
     //saving the changes
     await task.save();
 
-    return res.status(200).json({message: "Task Updated Successfully", success: true, task})
+    return res
+      .status(200)
+      .json({ message: "Task Updated Successfully", success: true, task });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -86,14 +142,31 @@ export const deleteTask = async (req, res) => {
   try {
     //taking taskId from params
     const { id } = req.params;
+    const { userId } = req.params;
+
+    const user = await User.findById(userId);
+    if (!user) {
+      return res
+        .status(400)
+        .json({ message: "User not Found", success: false });
+    }
+    
     //deleting Task
     const task = await Task.findByIdAndDelete(id);
-    //checking if task is present in DB or no
-    if(!task){
-      return res.json({message: "Task Not Found", success: false})
+     //checking if task is present in DB or no
+    if (!task) {
+      return res.json({ message: "Task Not Found", success: false });
     }
 
-    return res.status(200).json({message: "Task Deleted Successfully", success: true, task})
+    if (task.user != userId) {
+      return res
+        .status(401)
+        .json({ message: "Task Not Found", success: false });
+    }
+   
+    return res
+      .status(200)
+      .json({ message: "Task Deleted Successfully", success: true, task });
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
